requirements.txt
# -------------------------
# Core Python utilities
# -------------------------
python-dotenv>=1.0.1
pydantic>=2.7.0
typing-extensions>=4.10.0

# -------------------------
# Data handling (CRITICAL)
# -------------------------
pandas>=2.2.2
numpy>=1.26.4
openpyxl>=3.1.2

# -------------------------
# LangChain (LATEST SPLIT)
# -------------------------
langchain>=0.2.10
langchain-core>=0.2.18
langchain-community>=0.2.10
langchain-text-splitters>=0.2.2

# -------------------------
# Vector Store & Embeddings
# -------------------------
faiss-cpu>=1.8.0
tiktoken>=0.7.0

# -------------------------
# LLM Providers
# -------------------------
google-generativeai>=0.7.2

# -------------------------
# Utilities
# -------------------------
rich>=13.7.1
loguru>=0.7.2

# -------------------------
# Testing
# -------------------------
pytest>=8.2.0
